+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/reflection/RecordProbe.java
@@ -8,8 +8,6 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import java.util.stream.StreamSupport;
 
-import nl.jqno.equalsverifier.internal.util.PrimitiveMappers;
-
 public class RecordProbe<T> {
 
     private final Class<T> type;
@@ -37,37 +35,17 @@ public class RecordProbe<T> {
     }
 
     private String buildMessage(Throwable e, List<?> params) {
-        String prefix = "Record: failed to run constructor for record type " + type.getName()
+        String msg = "Record: failed to run constructor for record type " + type.getName()
                 + "\n   These were the values passed to the constructor: " + params;
 
         if (e.getCause() instanceof NullPointerException) {
-            return prefix + "\n   If the record does not accept null values for its constructor parameters,"
+            return msg + "\n   If the record does not accept null values for its constructor parameters,"
                     + " consider suppressing Warning.NULL_FIELDS.";
         }
 
-        boolean hasZeros = false;
-        boolean hasSomethingElse = false;
-        for (Object p : params) {
-            if (PrimitiveMappers.ZEROS.contains(p)) {
-                hasZeros = true;
-            }
-            else {
-                // nulls are already eliminated
-                hasSomethingElse = true;
-            }
-        }
-
-        String msg = prefix;
-        if (hasZeros) {
-            msg += """
-                   \n   If the record does not accept 0 or false as a value for its constructor parameters,\
-                   consider suppressing Warning.ZERO_FIELDS.""";
-        }
-        if (hasSomethingElse) {
-            msg += """
-                   \n   If the record does not accept any of the given values for its constructor parameters,\
-                   consider providing prefab values for the types of those fields.""";
-        }
+        msg += """
+               \n   If the record does not accept any of the given values for its constructor parameters, \
+               consider providing prefab values for the types of those fields.""";
         return msg;
     }
 }