+++ equalsverifier-core/src/main/java/nl/jqno/equalsverifier/internal/util/PrimitiveMappers.java
@@ -1,6 +1,5 @@
 package nl.jqno.equalsverifier.internal.util;
 
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,7 +11,6 @@ public final class PrimitiveMappers {
     public static final Map<Class<?>, Class<?>> PRIMITIVE_OBJECT_MAPPER = createPrimitiveObjectMapper();
     public static final Map<Class<?>, Object> DEFAULT_VALUE_MAPPER = createDefaultValueMapper(false);
     public static final Map<Class<?>, Object> DEFAULT_WRAPPED_VALUE_MAPPER = createDefaultValueMapper(true);
-    public static final Collection<Object> ZEROS = DEFAULT_WRAPPED_VALUE_MAPPER.values();
 
     private static Map<Class<?>, Class<?>> createPrimitiveObjectMapper() {
         var result = new HashMap<Class<?>, Class<?>>();